pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.35.0",
]
